'Un-disabled' a couple of MSVC warnings and fixed some occurances   of VC warnings in the source.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@15803 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/perftest/ptests2d.cpp
@@ -108,7 +108,7 @@ void StringTester::Setup (iGraphics3D* g3d, PerfTest* /*perftest*/)
     length = (int)screen_len;
     line = new char[length];
     char *buf = line;
-    int buf_len = length;
+    size_t buf_len = length;
     while (buf_len > tl_len)
     {
       strncpy (buf, tl, tl_len);
